-NSInteger is 64-bit on x86_64 so use long instead (cannot test for now, use previous revisions if broken).
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@63552 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/PcsxPlugin.h
@@ -29,7 +29,7 @@
 - (NSString *)getDisplayVersion;
 - (int)getType;
 - (NSString *)path;
-- (unsigned int)hash;
+- (unsigned long)hash;
 - (NSString *)description;
 - (BOOL)hasAboutAs:(int)type;
 - (BOOL)hasConfigureAs:(int)type;
